% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_calibrarion.R
\name{get.internal.cal}
\alias{get.internal.cal}
\title{Internal Calibration of Cox PH model}
\usage{
get.internal.cal(
  formula,
  TR.data,
  time,
  fig.name,
  fig.title,
  fig.width = 800,
  fig.height = 700,
  quality = 300,
  res = 100,
  B.num = 150,
  seed = 2019
)
}
\arguments{
\item{formula}{a formula used to specify a Cox PH model, see cph function.}

\item{TR.data}{the name of a data frame containing all needed variables, used as training dataset.}

\item{time}{a single value to specify the timepoint you are interested in for calibration.}

\item{fig.name}{a string for the name of the calibration plot.}

\item{fig.title}{a string for the title of the figure}

\item{fig.width}{a number of width of the calibration plot. Defaults to 800}

\item{fig.height}{a number of height of the calibration plot. Defaults to 700}

\item{quality}{the ‘quality’ of the JPEG image, as a percentage. Smaller values will give more compression but also more degradation of the image. Defaults to 300}

\item{res}{The nominal resolution in ppi which will be recorded in the bitmap file, is a positive integer. Defaults to 100}

\item{B.num}{a number of repetitions for bootstrap. Defaults to 150}

\item{seed}{a single value, interpreted as an integer, or NULL}
}
\description{
This function allows you to generate an internal calibration plot of a Cox PH model.
}
\details{
For other parameters in rms::calibrate, use default setting.

In cph function, surv=TRUE and time.inc=u were specified to make a calibration plot.
}
\examples{
n <- 500
set.seed(2019)
var1 <- 30 + 5*rnorm(n)
var2 <- factor(sample(c("White","Black"),n,rep=TRUE,prob=c(.3,.7)))
hazard <- .1*exp(.2*(var1-30)+.5*(var2=="Black"))
time_to_event <- -log(runif(n))/hazard
cens <- 15*runif(n)
censor <- ifelse(time_to_event <= cens,1,0)
TR <- data.frame(time_to_event = time_to_event,censor=censor, var1 = var1, var2= var2)
get.internal.cal(formula = Surv(time_to_event,censor) ~ var1 + var2,
                  TR.data = TR,time = 12,fig.name="name", fig.title='Survival_Plot')
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{nomovalidation}
