% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_validation.R
\name{get.internal.valid}
\alias{get.internal.valid}
\title{Internal Validation of Cox PH model}
\usage{
get.internal.valid(formula, TR.data, B.num = 200, seed = 2019)
}
\arguments{
\item{formula}{a formula used to specify a Cox PH model, see cph function.}

\item{TR.data}{the name of a data frame containing all needed variables, used as training dataset.}

\item{B.num}{a number of repetitions for bootstrap. Defaults to 150}

\item{seed}{a single value, interpreted as an integer, or NULL}
}
\description{
This function allows you to generate a Harrell's C-index of internal validation for a Cox PH model.
}
\details{
Returns C.index, a value of Harrell's C-index. C index is calculated from Dxy.

The relation between Dxy and C-index is Dxy=2*(C-0.5).
}
\examples{
n <- 500
set.seed(2019)
var1 <- 30 + 5*rnorm(n)
var2 <- factor(sample(c("White","Black"),n,rep=TRUE,prob=c(.3,.7)))
hazard <- .1*exp(.2*(var1-30)+.5*(var2=="Black"))
time_to_event <- -log(runif(n))/hazard
cens <- 15*runif(n)
censor <- ifelse(time_to_event <= cens,1,0)
TR <- data.frame(time_to_event = time_to_event,censor=censor, var1 = var1, var2= var2)
get.internal.valid(formula = Surv(time_to_event,censor) ~ var1 + var2,TR,B.num=2)
}
\author{
Xi "Ada" Wang

Ph.D. Student of Biostatistics

Penn State College of Medicine

xzw149@psu.edu
}
\keyword{nomovalidation}
